requirements.txt
annotated-types==0.7.0
asgiref==3.11.1
click==8.3.2
Django==6.0.4
django-ninja==1.6.2
gunicorn==25.3.0
h11==0.16.0
packaging==26.1
psycopg==3.3.3
psycopg-binary==3.3.3
pydantic==2.13.3
pydantic_core==2.46.3
python-dotenv==1.2.2
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
